修复了无法获取会话列表的问题

+++ oa-ai-service/src/main/java/com/personaowl/oa/ai/domain/vo/PageResultVO.java
@@ -2,5 +2,5 @@ package com.personaowl.oa.ai.domain.vo;
 
 import java.util.List;
 
-public record PageResultVO<T>(List<T> list, Integer page, Integer size, Long total) {
+public record PageResultVO<T>(List<T> list, Integer page, Integer size, Integer total) {
 }
